+++ find_work/cli/execute.py
@@ -4,23 +4,15 @@
 
 """ CLI subcommand for executing custom aliases. """
 
-import tomllib
 from collections.abc import Callable
 from importlib import import_module
-from importlib.resources import files
-from pathlib import Path
 from typing import Any
 
 import click
-import gentoopm
 from click_aliases import ClickAliasedGroup
-from deepmerge import always_merger
-from platformdirs import PlatformDirs
 
-import find_work.data
 from find_work.cli import Options
-from find_work.config import Config, ConfigAlias, ConfigModuleOption
-from find_work.constants import DEFAULT_CONFIG, ENTITY, PACKAGE
+from find_work.config import ConfigAlias, ConfigModuleOption, load_config
 from find_work.types import CliOptionKind
 
 
@@ -76,21 +68,5 @@ def load_aliases(group: ClickAliasedGroup) -> None:
     :param group: click group for new commands
     """
 
-    default_config = files(find_work.data).joinpath(DEFAULT_CONFIG).read_text()
-    toml = tomllib.loads(default_config)
-
-    pm = gentoopm.get_package_manager()
-    system_config = Path(pm.root) / "etc" / PACKAGE / "config.toml"
-    if system_config.is_file():
-        with open(system_config, "rb") as file:
-            always_merger.merge(toml, tomllib.load(file))
-
-    dirs = PlatformDirs(PACKAGE, ENTITY, roaming=True)
-    user_config = dirs.user_config_path / "config.toml"
-    if user_config.is_file():
-        with open(user_config, "rb") as file:
-            always_merger.merge(toml, tomllib.load(file))
-
-    config = Config(toml)
-    for alias in config.aliases:
+    for alias in load_config().aliases:
         group.command(aliases=alias.shortcuts)(_callback_from_config(alias))