+++ find_work/config.py
@@ -4,10 +4,19 @@
 
 """ Configuration parsing and validation. """
 
+import tomllib
 from abc import ABC
-from functools import cached_property
+from functools import cache, cached_property
+from importlib.resources import files
+from pathlib import Path
 from typing import Any
 
+import gentoopm
+from deepmerge import always_merger
+from platformdirs import PlatformDirs
+
+import find_work.data
+from find_work.constants import DEFAULT_CONFIG, ENTITY, PACKAGE
 from find_work.types import CliOptionKind
 
 
@@ -110,7 +119,7 @@ class ConfigAlias(ConfigBase):
 
         for opt_name, opt_value in opts.items():
             if isinstance(opt_value, dict) and len(opt_value) != 1:
-                raise self._value_error(f"module.{opt_name}", "invalid value")
+                raise self._value_error(f"{module}.{opt_name}", "invalid value")
             result.append(ConfigModuleOption(module, opt_name, opt_value,
                                              context=self._context + "options"))
         return result
@@ -140,6 +149,45 @@ class ConfigAlias(ConfigBase):
         return self._get("shortcuts", list, default=[])
 
 
+class ConfigFlag(ConfigBase):
+    """ Basic custom global flags parser and validator. """
+
+    def __init__(self, name: str, params: dict, *, context: str = ""):
+        self.name: str = name
+
+        self._obj: dict = params
+        self._context: str = f"{context}.{name}."
+
+    @cached_property
+    def params(self) -> dict[str, Any]:
+        """
+        Load new global flag's module options from the configuration.
+        """
+
+        result: dict[str, Any] = {}
+        opts = self._get("params", dict)
+
+        for opt_name, opt_value in opts.items():
+            result[opt_name] = opt_value
+        return result
+
+    @cached_property
+    def description(self) -> str:
+        """
+        Load new global option's help text from the configuration.
+        """
+
+        return self._get("description", str)
+
+    @cached_property
+    def shortcuts(self) -> list[str]:
+        """
+        Load new global option's aliases from the configuration.
+        """
+
+        return self._get("shortcuts", list, default=[])
+
+
 class Config(ConfigBase):
     """ Basic configuration parser and validator. """
 
@@ -158,3 +206,39 @@ class Config(ConfigBase):
             result.append(ConfigAlias(name, params,
                                       context=self._context + "alias"))
         return result
+
+    @cached_property
+    def flags(self) -> list[ConfigFlag]:
+        """
+        Load custom flags from the configuration.
+        """
+
+        result: list[ConfigFlag] = []
+        for name, params in self._get("flag", dict, default={}).items():
+            result.append(ConfigFlag(name, params,
+                                     context=self._context + "flag"))
+        return result
+
+
+@cache
+def load_config() -> Config:
+    """
+    Load configuration files.
+    """
+
+    default_config = files(find_work.data).joinpath(DEFAULT_CONFIG).read_text()
+    toml = tomllib.loads(default_config)
+
+    pm = gentoopm.get_package_manager()
+    system_config = Path(pm.root) / "etc" / PACKAGE / "config.toml"
+    if system_config.is_file():
+        with open(system_config, "rb") as file:
+            always_merger.merge(toml, tomllib.load(file))
+
+    dirs = PlatformDirs(PACKAGE, ENTITY, roaming=True)
+    user_config = dirs.user_config_path / "config.toml"
+    if user_config.is_file():
+        with open(user_config, "rb") as file:
+            always_merger.merge(toml, tomllib.load(file))
+
+    return Config(toml)